requirements.txt
nltk
sentence-transformers
faiss-cpu
google-generativeai
python-dotenv
